SERVER: unify http code with client
