+++ code/client/CMakeLists.txt
@@ -68,14 +68,20 @@ if (WIN32)
 	list(APPEND SRCS ../sys/sys_win32.c)
 	list(APPEND SRCS ../sys/con_passive.c)
 	list(APPEND SRCS ../sys/win_resource.rc)
-	list(APPEND SRCS cl_http_windows.c)
+	if (USE_HTTP)
+		list(APPEND SRCS cl_http_windows.c)
+		list(APPEND SRCS ../server/sv_http_windows.c)
+	endif()
 else()
 	if(APPLE)
 		list(APPEND SRCS ../sys/sys_osx.m)
 	endif()
 	list(APPEND SRCS ../sys/sys_unix.c)
 	list(APPEND SRCS ../sys/con_tty.c)
-	list(APPEND SRCS cl_http_curl.c)
+	if (USE_HTTP)
+		list(APPEND SRCS cl_http_curl.c)
+		list(APPEND SRCS ../server/sv_http_curl.c)
+	endif()
 endif()
 
 if (USE_MUMBLE)
